Removal of ES dependency in arlas-rest
