+++ arlas-rest/src/main/java/io/arlas/server/rest/explore/search/SearchRESTService.java
@@ -23,31 +23,30 @@ import com.codahale.metrics.annotation.Timed;
 import io.arlas.server.app.Documentation;
 import io.arlas.server.exceptions.ArlasException;
 import io.arlas.server.model.CollectionReference;
-import io.arlas.server.model.Link;
 import io.arlas.server.model.request.MixedRequest;
-import io.arlas.server.model.request.Page;
 import io.arlas.server.model.request.Search;
 import io.arlas.server.model.response.Error;
-import io.arlas.server.model.response.*;
+import io.arlas.server.model.response.Hits;
 import io.arlas.server.ns.ATOM;
 import io.arlas.server.rest.explore.ExploreRESTServices;
 import io.arlas.server.services.ExploreServices;
-import io.arlas.server.utils.*;
+import io.arlas.server.utils.CheckParams;
+import io.arlas.server.utils.ColumnFilterUtil;
+import io.arlas.server.utils.ParamsParser;
 import io.dropwizard.jersey.params.IntParam;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import org.apache.commons.lang.BooleanUtils;
-import org.elasticsearch.search.SearchHit;
-import org.elasticsearch.search.SearchHits;
+
 import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
 import java.io.IOException;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.List;
+import java.util.Optional;
 
 public class SearchRESTService extends ExploreRESTServices {
 
@@ -210,7 +209,7 @@ public class SearchRESTService extends ExploreRESTServices {
         request.headerRequest = searchHeader;
         request.columnFilter = ColumnFilterUtil.getCollectionRelatedColumnFilter(columnFilter, collectionReference);;
 
-        Hits hits = getArlasHits(request, collectionReference, BooleanUtils.isTrue(flat), uriInfo,"GET");
+        Hits hits = this.exploreServices.getSearchHits(request, collectionReference, BooleanUtils.isTrue(flat), uriInfo,"GET");
         return cache(Response.ok(hits), maxagecache);
     }
 
@@ -288,147 +287,7 @@ public class SearchRESTService extends ExploreRESTServices {
         request.headerRequest = searchHeader;
         request.columnFilter = ColumnFilterUtil.getCollectionRelatedColumnFilter(columnFilter, collectionReference);
 
-        Hits hits = getArlasHits(request, collectionReference, (search.form != null && BooleanUtils.isTrue(search.form.flat)),uriInfo,"POST");
+        Hits hits = this.exploreServices.getSearchHits(request, collectionReference, (search.form != null && BooleanUtils.isTrue(search.form.flat)),uriInfo,"POST");
         return cache(Response.ok(hits), maxagecache);
-    }
-
-
-    protected Hits getArlasHits(MixedRequest request, CollectionReference collectionReference, Boolean flat, UriInfo uriInfo, String method) throws ArlasException, IOException {
-        SearchHits searchHits = this.getExploreServices().search(request, collectionReference);
-        Search searchRequest  = (Search)request.basicRequest;
-        Hits hits = new Hits(collectionReference.collectionName);
-        hits.totalnb = searchHits.getTotalHits().value;
-        hits.nbhits = searchHits.getHits().length;
-        HashMap<String,Link> links = new HashMap<>();
-        hits.hits = new ArrayList<>((int) hits.nbhits);
-        List<SearchHit>searchHitList= Arrays.asList(searchHits.getHits());
-        if(searchRequest.page != null && searchRequest.page.before != null ){
-            Collections.reverse(searchHitList);
-        }
-        for (SearchHit hit : searchHitList) {
-            hits.hits.add(new Hit(collectionReference, hit.getSourceAsMap(), searchRequest.returned_geometries, flat, false));
-        }
-        Link self = new Link();
-        self.href = getRequestUri(uriInfo);
-        self.method = method;
-        Link next = null;
-        Link previous = null;
-        int lastIndex = (int) hits.nbhits -1;
-        String sortParam = searchRequest.page != null ? searchRequest.page.sort : null;
-        String afterParam = searchRequest.page != null ? searchRequest.page.after : null;
-        String beforeParam = searchRequest.page != null ? searchRequest.page.before : null;
-        Integer sizeParam = searchRequest.page != null ? searchRequest.page.size : ExploreServices.SEARCH_DEFAULT_PAGE_SIZE;
-        String lastHitAfter = "";
-        String firstHitAfter = "";
-        if (lastIndex >= 0 && sizeParam == hits.nbhits && sortParam != null && (afterParam != null || sortParam.contains(collectionReference.params.idPath))) {
-            next = new Link();
-            next.method = method;
-            // Use sorted value of last element return by ES to build after param of next & previous link
-            lastHitAfter = Arrays.stream(searchHitList.get(lastIndex).getSortValues()).map(value->value.toString()).collect(Collectors.joining(","));
-        }
-        if (searchHitList.size()>0 && sortParam != null && (beforeParam != null || sortParam.contains(collectionReference.params.idPath))) {
-            previous = new Link();
-            previous.method = method;
-            firstHitAfter = Arrays.stream(searchHitList.get(0).getSortValues()).map(value->value.toString()).collect(Collectors.joining(","));
-        }
-
-        switch (method){
-            case"GET":
-                links.put("self", self);
-                if (next != null){
-                    next.href = getNextHref(uriInfo, lastHitAfter);
-                    links.put("next", next);
-                }
-                if (previous != null){
-                    previous.href = getPreviousHref(uriInfo, firstHitAfter);
-                    links.put("previous", previous);
-                }
-                break;
-            case"POST":
-                self.body = searchRequest;
-                links.put("self", self);
-                if (next != null){
-                    Page nextPage = new Page();
-                    Search search = new Search();
-                    search.filter = searchRequest.filter;
-                    search.form = searchRequest.form;
-                    search.projection =searchRequest.projection;
-                    search.returned_geometries = searchRequest.returned_geometries;
-                    nextPage.sort=searchRequest.page.sort;
-                    nextPage.size=searchRequest.page.size;
-                    nextPage.from =searchRequest.page.from;
-                    nextPage.after = lastHitAfter;
-                    search.page = nextPage;
-                    next.href = self.href;
-                    next.body = search;
-                    links.put("next", next);
-                }
-                if (previous != null){
-                    Page previousPage = new Page();
-                    Search search = new Search();
-                    search.filter = searchRequest.filter;
-                    search.form = searchRequest.form;
-                    search.projection =searchRequest.projection;
-                    search.returned_geometries = searchRequest.returned_geometries;
-                    previousPage.sort=searchRequest.page.sort;
-                    previousPage.size=searchRequest.page.size;
-                    previousPage.from =searchRequest.page.from;
-                    previousPage.before = firstHitAfter;
-                    search.page = previousPage;
-                    previous.href = self.href;
-                    previous.body = search;
-                    links.put("previous", previous);
-                }
-                break;
-        }
-        hits.links = links;
-        return hits;
-    }
-
-    private String getBaseUri(UriInfo uriInfo) {
-        String baseUri = this.exploreServices.getBaseUri();
-        if (StringUtil.isNullOrEmpty(baseUri)) {
-            baseUri = uriInfo.getBaseUri().toString();
-        }
-        return baseUri;
-    }
-
-    private String getPathUri(UriInfo uriInfo) {
-        return uriInfo.getPath();
-    }
-
-    private String getAbsoluteUri(UriInfo uriInfo) {
-        return getBaseUri(uriInfo) + getPathUri(uriInfo);
-    }
-
-    private String getQueryParameters(UriInfo uriInfo) {
-        return uriInfo.getRequestUriBuilder().toTemplate().replace(uriInfo.getAbsolutePath().toString(), "");
-    }
-
-    private String getNextQueryParameters(UriInfo uriInfo, String afterValue) {
-        return uriInfo.getRequestUriBuilder()
-                .replaceQueryParam("after", afterValue)
-                .replaceQueryParam("before", "")
-                .toTemplate()
-                .replace(uriInfo.getAbsolutePath().toString(), "").replace("&before=","").replace("before=&","");
-    }
-
-    private String getPreviousQueryParameters(UriInfo uriInfo, String afterValue) {
-        return uriInfo.getRequestUriBuilder()
-                .replaceQueryParam("after", "")
-                .replaceQueryParam("before", afterValue)
-                .toTemplate()
-                .replace(uriInfo.getAbsolutePath().toString(), "").replace("&after=","").replace("after=&","");
-    }
-
-    private String getRequestUri(UriInfo uriInfo) {
-        return getAbsoluteUri(uriInfo) + getQueryParameters(uriInfo);
-    }
-
-    private String getNextHref(UriInfo uriInfo, String afterValue) {
-        return getAbsoluteUri(uriInfo) + getNextQueryParameters(uriInfo, afterValue);
-    }
-    private String getPreviousHref(UriInfo uriInfo, String afterValue) {
-        return getAbsoluteUri(uriInfo) + getPreviousQueryParameters(uriInfo, afterValue);
     }
 }
